Threads created and tests started
Printing is now beautiful

Theres too much delay when trying to print the status of the philosophers at the same time. Should that be a problem? Maybe is because of my printing function and its various printfs. Or maybe I could use some variable to be more precise or something. Not really sure if that is correct