Consistency is finally achieved.
There is no need for a monitor

There is an error when trying to initialize a lot of threads (like 40000). I manage that error but there doesnt seem to be a way to avoid it. It is a fsnaitizer=address type error, but I dont think its my fault because I correctly exit right before the error is detected